Adjusted data types for inst and mtl ids from uint32_t to uint16_t. Subx/y to uint8_t. No mtl override at 0xffff instead of -1.

+++ src/rend/renderer.c
@@ -94,7 +94,7 @@ void renderer_update(scene *s, bool converge)
   }
 
   if(s->dirty & RT_TRI) {
-    for(uint32_t i=0; i<s->mesh_cnt; i++) {
+    for(uint16_t i=0; i<s->mesh_cnt; i++) {
       mesh *m = &s->meshes[i];
       gpu_write_buf(BT_TRI, m->ofs * sizeof(*m->tris), m->tris, m->tri_cnt * sizeof(*m->tris));
       gpu_write_buf(BT_NRM, m->ofs * sizeof(*m->tri_nrms), m->tri_nrms, m->tri_cnt * sizeof(*m->tri_nrms));
